Added tests for URL and Token

+++ ursh/blueprints/api/resources.py
@@ -1,7 +1,7 @@
 import re
 from uuid import UUID
 
-from flask import Response, g
+from flask import Response, g, jsonify
 from flask_apispec import MethodResource, marshal_with, use_kwargs
 from werkzeug.exceptions import BadRequest, Conflict, MethodNotAllowed, NotFound
 
@@ -21,7 +21,7 @@ class TokenResource(MethodResource):
         if Token.query.filter_by(name=kwargs['name']).count() != 0:
             raise Conflict({'message': 'Token with name exists', 'args': ['name']})
         new_token = Token()
-        if not validate_callback_url(kwargs.get('callback_url')):
+        if not None and not validate_url(kwargs.get('callback_url')):
             raise generate_bad_request('missing-args', 'Callback URL is invalid', args=['callback_url'])
         populate_from_dict(new_token, kwargs, ('name', 'is_admin', 'is_blocked', 'callback_url'))
         db.session.add(new_token)
@@ -41,7 +41,7 @@ class TokenResource(MethodResource):
         token = Token.query.filter_by(api_key=api_key).one_or_none()
         if token is None:
             raise NotFound({'message': 'API key does not exist', 'args': ['api_key']})
-        if not validate_callback_url(kwargs.get('callback_url')):
+        if not None and not validate_url(kwargs.get('callback_url')):
             raise generate_bad_request('missing-args', 'Callback URL is invalid', args=['callback_url'])
         populate_from_dict(token, kwargs, ('is_admin', 'is_blocked', 'callback_url'))
         db.session.commit()
@@ -84,7 +84,8 @@ class URLResource(MethodResource):
     @marshal_with(URLSchema(strict=True), code=201)
     @use_kwargs(URLSchema)
     def post(self, **kwargs):
-        if kwargs.get('url') is None:
+        posted_url = kwargs.get('url')
+        if posted_url is None or not validate_url(posted_url):
             raise generate_bad_request('missing-args', 'URL invalid or missing',
                                        args=['url'])
         if kwargs['shortcut']:
@@ -99,19 +100,19 @@ class URLResource(MethodResource):
     @authorize_request_for_url
     @marshal_with(URLSchema(strict=True), code=201)
     def put(self, shortcut, **kwargs):
-        if not kwargs.get('url'):
+        posted_url = kwargs.get('url')
+        if posted_url is None or not validate_url(posted_url):
             raise generate_bad_request('missing-args', 'URL invalid or missing',
                                        args=['url'])
         existing_url = URL.query.filter_by(shortcut=shortcut).one_or_none()
         if existing_url:
             if kwargs.get('allow_reuse'):
-                return existing_url
+                return existing_url, 201
             else:
                 raise Conflict({'message': 'Shortcut already exists', 'args': ['shortcut']})
         new_url = create_new_url(data=kwargs, shortcut=shortcut)
         db.session.add(new_url)
         db.session.commit()
-
         return new_url, 201
 
     @use_kwargs(URLSchema)
@@ -124,35 +125,47 @@ class URLResource(MethodResource):
         url = URL.query.filter_by(shortcut=shortcut).one_or_none()
         if not url:
             raise NotFound({'message': 'Shortcut does not exist', 'args': ['shortcut']})
+        posted_url = kwargs.get('url')
+        if posted_url is None or not validate_url(posted_url):
+            raise generate_bad_request('missing-args', 'URL invalid or missing',
+                                       args=['url'])
         url.custom_data = metadata
         populate_from_dict(url, kwargs, ('shortcut', 'url', 'allow_reuse'))
         db.session.commit()
         return url
 
-    @use_kwargs(URLSchema)
     @authorize_request_for_url
+    @use_kwargs(URLSchema)
     def delete(self, shortcut, **kwargs):
         if not shortcut:
             raise MethodNotAllowed
-        url = URL.query.filter_by(shortcut=shortcut).first_or_404()
+        url = URL.query.filter_by(shortcut=shortcut).one_or_none()
+        if not url:
+            raise NotFound({'message': 'Shortcut does not exist', 'args': ['shortcut']})
         db.session.delete(url)
         db.session.commit()
         return Response(status=204)
 
+    @marshal_many_or_one(URLSchema, 'shortcut', code=200)
     @use_kwargs(URLSchema)
     @authorize_request_for_url
-    @marshal_many_or_one(URLSchema, 'shortcut')
-    def get(self, shortcut=None, **kwargs):
+    def get(self, shortcut, **kwargs):
         if shortcut is None:
             metadata = kwargs.get('metadata')
             filters = []
+            if kwargs.get('url'):
+                filters.append(URL.url == kwargs.get('url'))
             if not g.token.is_admin or not kwargs.get('all'):
-                filters = [URL.token == g.token]
+                filters.append(URL.token == g.token)
             if metadata:
                 filters.append(URL.custom_data.contains(metadata))
             return URL.query.filter(*filters).all()
         else:
-            return URL.query.filter_by(shortcut=shortcut).first_or_404()
+            url = URL.query.filter_by(shortcut=shortcut).one_or_none()
+            if not url:
+                raise NotFound({'message': 'Shortcut does not exist', 'args': ['shortcut']})
+            else:
+                return url
 
 
 def populate_from_dict(obj, values, fields):
@@ -177,5 +190,5 @@ def generate_bad_request(error_code, message, **kwargs):
     return BadRequest(message_dict)
 
 
-def validate_callback_url(url, _re=re.compile(r'^https?://[^/:]+(:[0-9]+)?(/.*)?$')):
-    return url is None or _re.match(url) is not None
+def validate_url(url, _re=re.compile(r'^https?://[^/:]+(:[0-9]+)?(/.*)?$')):
+    return _re.match(url) is not None
